Adds value cloning to prevent recursive  v-model updates
