+++ packages/core/__tests__/lists.spec.ts
@@ -7,7 +7,7 @@ import {
   createPlaceholder,
 } from '../src'
 import { clone } from '@formkit/utils'
-import { createNameTree } from '.jest/helpers'
+import { createNameTree } from '.tests/helpers'
 
 describe('lists', () => {
   it('can get a node’s index', () => {
